backend/Audio/requirements.txt
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
letta-client>=0.0.1
openai>=1.56.0
pydantic>=2.11.1,<3
beautifulsoup4==4.13.4
bs4==0.0.2
tiktoken==0.9.0
transformers==4.51.3
PyPDF2==3.0.1
PyYAML==6.0.2
requests==2.32.3
rich<14,>=13.7.1
sentence-transformers==4.1.0
numpy<2